requirements.txt
Flask
pandas
scikit-learn
joblib
python-dotenv
openai
requests
